refactor(TemplateMatch): 把导入面板从弹窗壳里抽出来
内容与外壳分开：ImportPanel 负责 Tab、勾选、就地落库，ImportModal 只剩弹窗框。
数据源保存后的引导行为不变。

抽出来是因为采集配置那条流程也要用它，但形态不同：那里组件、业务组、数据源
在上下文里都已确定，没有可挑的，再叠一层弹窗只是多一次点击，直接内联面板即可。

顺带把调用方专属的副作用挪出面板：markDsJourney 记的是数据源的旅程，采集流程
复用时不该写它。改由各自外层在 onImported 里做，面板不认识这些。

新增 defaultBgid：调用方知道模板该落哪个业务组时预填（如采集配置所在的组），
不传则沿用「只有一个业务组就自动选中」的老规则，用户仍可改。
